refactor: rename snapshotx to snapshot-x
