+++ apps/ui/src/components/Layout/App.vue
@@ -64,7 +64,11 @@ const hasSwipeableContent = computed(() => hasSidebar.value || hasAppNav.value);
 
 const hasPlaceHolderSidebar = computed(
   () =>
-    !['space-proposal', 'create'].includes(String(route.matched[0]?.name)) &&
+    ![
+      'space-proposal',
+      'space-create-snapshot',
+      'space-create-snapshot-x'
+    ].includes(String(route.matched[0]?.name)) &&
     !['space-editor', 'space-proposal'].includes(String(route.matched[1]?.name))
 );
