+++ apps/ui/src/networks/types.ts
@@ -423,7 +423,7 @@ export type ReadWriteNetwork = BaseNetwork & {
 };
 export type Network = ReadOnlyNetwork | ReadWriteNetwork;
 
-export type ExplorePageProtocol = 'snapshot' | 'snapshotx';
+export type ExplorePageProtocol = 'snapshot' | 'snapshot-x';
 
 export type ProtocolConfig = {
   key: ExplorePageProtocol;
