+++ apps/ui/src/routes/default.ts
@@ -1,5 +1,4 @@
 import App from '@/views/App.vue';
-import Create from '@/views/Create.vue';
 import Ecosystem from '@/views/Ecosystem.vue';
 import Landing from '@/views/Landing.vue';
 import Explore from '@/views/My/Explore.vue';
@@ -12,6 +11,8 @@ import Contacts from '@/views/Settings/Contacts.vue';
 import SettingsSpaces from '@/views/Settings/Spaces.vue';
 import Settings from '@/views/Settings.vue';
 import Site from '@/views/Site.vue';
+import SpaceCreateSnapshot from '@/views/Space/CreateSnapshot.vue';
+import SpaceCreateSnapshotX from '@/views/Space/CreateSnapshotX.vue';
 import Space from '@/views/Space.vue';
 import Terms from '@/views/Terms.vue';
 import User from '@/views/User.vue';
@@ -40,9 +41,18 @@ export default [
 
   {
     path: '/create',
-    redirect: { name: 'create', params: { protocol: 'snapshot-x' } }
+    redirect: { name: 'create-snapshot-x' }
   },
-  { path: '/create/:protocol', name: 'create', component: Create },
+  {
+    path: '/create/snapshot-x',
+    name: 'space-create-snapshot-x',
+    component: SpaceCreateSnapshotX
+  },
+  {
+    path: '/create/snapshot',
+    name: 'space-create-snapshot',
+    component: SpaceCreateSnapshot
+  },
   {
     path: '/settings',
     name: 'settings',
