+++ apps/ui/src/views/Space/CreateSnapshot.vue
@@ -2,7 +2,7 @@
 import { StepRecords } from '@/components/Ui/Stepper.vue';
 import { CHAIN_IDS } from '@/helpers/constants';
 import { clone } from '@/helpers/utils';
-import { getNetwork } from '@/networks';
+import { getNetwork, metadataNetwork } from '@/networks';
 import { StrategyConfig } from '@/networks/types';
 import { Member, NetworkID, Validation } from '@/types';
 
@@ -30,6 +30,8 @@ const DEFAULT_STEP_ERRORS = {
   profile: false,
   voting: false
 };
+
+const networkId: NetworkID = metadataNetwork;
 
 type extendedStepRecords = Record<
   keyof StepRecords,
@@ -103,12 +105,11 @@ const STEPS: extendedStepRecords = {
   }
 } as const;
 
-const props = defineProps<{ networkId: NetworkID }>();
-
 const { createSpaceRaw } = useActions();
 const { web3, authInitiated } = useWeb3();
 const router = useRouter();
 const { modalAccountOpen } = useModal();
+useTitle('Create space');
 
 const sending = ref(false);
 const stepsErrors = ref<Record<string, boolean>>(clone(DEFAULT_STEP_ERRORS));
@@ -118,7 +119,7 @@ const settingsForm = ref(clone(DEFAULT_SETTINGS_FORM));
 const space = computed(() => ({
   turbo: false,
   verified: false,
-  network: props.networkId,
+  network: networkId,
   id: settingsForm.value.id,
   cover: settingsForm.value.cover,
   avatar: settingsForm.value.avatar
@@ -176,13 +177,13 @@ const formattedSpaceSettings = computed(() => {
 });
 
 async function handleSubmit() {
-  const compositeSpaceId = `${props.networkId}:${settingsForm.value.id}`;
+  const compositeSpaceId = `${networkId}:${settingsForm.value.id}`;
 
   try {
     sending.value = true;
 
     await createSpaceRaw(
-      props.networkId,
+      networkId,
       settingsForm.value.id,
       JSON.stringify(formattedSpaceSettings.value)
     );
@@ -224,108 +225,111 @@ watch(
 </script>
 
 <template>
-  <UiStepper :steps="STEPS" :submitting="sending" @submit="handleSubmit">
-    <template #content="{ currentStep, goToNext }">
-      <UiContainerSettings
-        :title="STEPS[currentStep].contentTitle"
-        :description="STEPS[currentStep].contentDescription"
-      >
-        <div v-if="currentStep === 'welcome'" class="space-y-4">
-          <div>
-            You will be guided through the process of creating a space.
-            <br />
-            Don't worry, all settings can be changed later.
-          </div>
-          <div class="border py-3 px-4 rounded-lg flex gap-3">
-            <IH-Book-open class="shrink-0 size-[30px]" />
+  <div class="pt-5 max-w-[50rem] mx-auto px-4">
+    <UiStepper :steps="STEPS" :submitting="sending" @submit="handleSubmit">
+      <template #content="{ currentStep, goToNext }">
+        <UiContainerSettings
+          :title="STEPS[currentStep].contentTitle"
+          :description="STEPS[currentStep].contentDescription"
+        >
+          <div v-if="currentStep === 'welcome'" class="space-y-4">
             <div>
-              <b>Not sure how to setup your space?</b>
+              You will be guided through the process of creating a space.
               <br />
-              Learn more in the
-              <AppLink
-                :to="'https://docs.snapshot.box/user-guides/spaces/create'"
-              >
-                documentation
-              </AppLink>
-              or contact support on
-              <AppLink :to="'https://help.snapshot.box/en'"> Helpdesk </AppLink
-              >.
+              Don't worry, all settings can be changed later.
+            </div>
+            <div class="border py-3 px-4 rounded-lg flex gap-3">
+              <IH-Book-open class="shrink-0 size-[30px]" />
+              <div>
+                <b>Not sure how to setup your space?</b>
+                <br />
+                Learn more in the
+                <AppLink
+                  :to="'https://docs.snapshot.box/user-guides/spaces/create'"
+                >
+                  documentation
+                </AppLink>
+                or contact support on
+                <AppLink :to="'https://help.snapshot.box/en'">
+                  Helpdesk </AppLink
+                >.
+              </div>
             </div>
           </div>
-        </div>
-        <EnsConfiguratorOffchain
-          v-if="currentStep === 'id'"
-          v-model="settingsForm.id"
-          :network-id="networkId"
-          @select="goToNext()"
-        />
-        <FormSpaceProfile
-          v-if="currentStep === 'profile'"
-          :form="settingsForm"
-          :space="space"
-          @errors="v => handleErrors('profile', v)"
-        />
-        <div v-if="currentStep === 'network'" class="s-box mb-4">
-          <UiSelectorNetwork
-            v-model="settingsForm.chainId"
-            :definition="{
-              type: 'number',
-              title: 'Network',
-              tooltip:
-                'Networks can also be specified in individual strategies, delegations, treasuries, etc...',
-              examples: ['Select network'],
-              networkId: networkId,
-              networksListKind: 'offchain'
-            }"
+          <EnsConfiguratorOffchain
+            v-if="currentStep === 'id'"
+            v-model="settingsForm.id"
+            :network-id="networkId"
+            @select="goToNext()"
           />
-        </div>
-        <SetupStrategiesConfiguratorOffchain
-          v-if="currentStep === 'strategies'"
-          v-model="settingsForm.strategies"
-          :snapshot-chain-id="settingsForm.chainId"
-          :network-id="networkId"
-          :space="space"
-        />
-        <ProposalValidationConfigurator
-          v-if="currentStep === 'proposal'"
-          v-model="settingsForm.proposalValidation"
-          :network-id="networkId"
-          :snapshot-chain-id="settingsForm.chainId"
-        />
-        <template v-if="currentStep === 'voting'">
-          <div class="mb-3">
-            The voting delay is the time interval between the creation of a
-            proposal and the start of voting. The voting period is the duration
-            for which the proposal remains open for voting.<br />
-            If these values are left empty, the proposal author will be able to
-            set them.
-          </div>
-          <FormVoting
+          <FormSpaceProfile
+            v-if="currentStep === 'profile'"
             :form="settingsForm"
-            :selected-network-id="networkId"
-            @errors="v => handleErrors('voting', v)"
+            :space="space"
+            @errors="v => handleErrors('profile', v)"
           />
-        </template>
-        <FormSpaceMembers
-          v-if="currentStep === 'members'"
-          v-model="settingsForm.members"
-          :network-id="networkId"
-          :is-controller="true"
-          :is-admin="true"
-        />
-        <template v-if="currentStep === 'controller'">
-          <UiMessage type="danger" class="mb-3">
-            The controller is set to the ENS name owner. Any change of the ENS
-            name ownership will also change the controller.</UiMessage
-          >
-          <FormSpaceController
-            :controller="web3.account"
-            :network="getNetwork(networkId)"
-            :disabled="true"
+          <div v-if="currentStep === 'network'" class="s-box mb-4">
+            <UiSelectorNetwork
+              v-model="settingsForm.chainId"
+              :definition="{
+                type: 'number',
+                title: 'Network',
+                tooltip:
+                  'Networks can also be specified in individual strategies, delegations, treasuries, etc...',
+                examples: ['Select network'],
+                networkId: networkId,
+                networksListKind: 'offchain'
+              }"
+            />
+          </div>
+          <SetupStrategiesConfiguratorOffchain
+            v-if="currentStep === 'strategies'"
+            v-model="settingsForm.strategies"
+            :snapshot-chain-id="settingsForm.chainId"
+            :network-id="networkId"
+            :space="space"
           />
-        </template>
-      </UiContainerSettings>
-    </template>
-    <template #submit-text> Create space</template>
-  </UiStepper>
+          <ProposalValidationConfigurator
+            v-if="currentStep === 'proposal'"
+            v-model="settingsForm.proposalValidation"
+            :network-id="networkId"
+            :snapshot-chain-id="settingsForm.chainId"
+          />
+          <template v-if="currentStep === 'voting'">
+            <div class="mb-3">
+              The voting delay is the time interval between the creation of a
+              proposal and the start of voting. The voting period is the
+              duration for which the proposal remains open for voting.<br />
+              If these values are left empty, the proposal author will be able
+              to set them.
+            </div>
+            <FormVoting
+              :form="settingsForm"
+              :selected-network-id="networkId"
+              @errors="v => handleErrors('voting', v)"
+            />
+          </template>
+          <FormSpaceMembers
+            v-if="currentStep === 'members'"
+            v-model="settingsForm.members"
+            :network-id="networkId"
+            :is-controller="true"
+            :is-admin="true"
+          />
+          <template v-if="currentStep === 'controller'">
+            <UiMessage type="danger" class="mb-3">
+              The controller is set to the ENS name owner. Any change of the ENS
+              name ownership will also change the controller.</UiMessage
+            >
+            <FormSpaceController
+              :controller="web3.account"
+              :network="getNetwork(networkId)"
+              :disabled="true"
+            />
+          </template>
+        </UiContainerSettings>
+      </template>
+      <template #submit-text> Create space</template>
+    </UiStepper>
+  </div>
 </template>
